feat: Add Silly Tavern demonstration blog post
This commit adds a new blog post titled "Silly Tavern demonstration – Persona/Voice clone" to showcase the Silly Tavern technology. The blog post includes a description of the project, its potential applications in cognitive science and psychology research, and the technologies used. It also includes a demonstration video and a screenshot of the Silly Tavern UI.

[Link to the blog post](../_posts/2024-03-19-SillyTavern-custom.md)