feat: Add MBTI Comparison App post
Add a new post titled "Exploring the MBTI Comparison App" that introduces a project I've been working on. The post includes a screenshot of the app and a description of its purpose. It also provides a link to the demo and invites readers to try it out and learn more about the project.